--- app/popover/index.ts
@@ -1,1 +1,0 @@
-export { Popover, type PopoverAlignment, type PopoverProps } from './Popover';
