+++ lib/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@moondreamsdev/dreamer-ui",
-  "version": "1.7.10-test.61",
+  "version": "1.7.10-test.62",
   "publishConfig": {
     "access": "public"
   },
