+++ lib/src/components/popover/Popover.tsx
@@ -1,7 +1,7 @@
-import React, { useEffect, useId, useRef } from 'react';
+import React, { useEffect, useId, useMemo, useRef, useState } from 'react';
 import { join, mergeRefs } from '../../utils';
-
-export type PopoverAlignment = 'left' | 'center' | 'right';
+import { useAutoSwitchPlacement } from './hooks';
+import { placementVariants, PopoverAlignment, PopoverPlacement } from './variants';
 
 export interface PopoverProps {
   id?: string;
@@ -9,17 +9,16 @@ export interface PopoverProps {
   isOpen?: boolean;
   children: React.ReactNode;
   trigger: React.ReactElement;
+  placement?: PopoverPlacement;
   alignment?: PopoverAlignment;
   className?: string;
   closeOnOverlayClick?: boolean;
   closeOnTriggerClick?: boolean;
+  /** Offset of the popover from the trigger element in pixel */
+  offset?: number;
+  /** Automatically switch placement to opposite side if there is not enough space in the viewport */
+  autoSwitchPlacement?: boolean;
 }
-
-const POPOVER_ALIGNMENT_CLASSES: Record<PopoverAlignment, string> = {
-  left: 'left-0 origin-top-left',
-  center: 'left-1/2 -translate-x-1/2 origin-top',
-  right: 'right-0 origin-top-right',
-};
 
 export function Popover({
   id,
@@ -29,14 +28,26 @@ export function Popover({
   className,
   closeOnOverlayClick = true,
   trigger,
+  placement = 'bottom',
   alignment = 'center',
   closeOnTriggerClick = true,
+  offset = 8,
+  autoSwitchPlacement = true,
 }: PopoverProps) {
-  const [internalIsOpen, setInternalIsOpen] = React.useState(isOpen !== undefined ? isOpen : false);
+  const [internalIsOpen, setInternalIsOpen] = useState(isOpen ?? false);
   const popoverRef = useRef<HTMLDivElement>(null);
   const triggerRef = useRef<HTMLElement>(null);
   const previousFocusRef = useRef<HTMLElement | null>(null);
   const popoverId = useId();
+
+  const effectivePlacement = useAutoSwitchPlacement({
+    internalIsOpen,
+    autoSwitchPlacement,
+    placement,
+    offset,
+    triggerRef,
+    popoverRef,
+  });
 
   useEffect(() => {
     if (isOpen !== undefined) {
@@ -47,14 +58,11 @@ export function Popover({
   // Handle keyboard navigation
   useEffect(() => {
     if (!internalIsOpen) return;
-
     const handleKeyDown = (event: KeyboardEvent) => {
       if (event.key === 'Escape') {
-        event.preventDefault();
         setInternalIsOpen(false);
       }
     };
-
     document.addEventListener('keydown', handleKeyDown);
     return () => document.removeEventListener('keydown', handleKeyDown);
   }, [internalIsOpen]);
@@ -62,10 +70,8 @@ export function Popover({
   // Handle click outside
   useEffect(() => {
     if (!internalIsOpen || !closeOnOverlayClick) return;
-
-    const handleAction = (event: PointerEvent | MouseEvent) => {
+    const handleClickOutside = (event: MouseEvent) => {
       const target = event.target as Node;
-
       if (
         popoverRef.current &&
         !popoverRef.current.contains(target) &&
@@ -75,64 +81,56 @@ export function Popover({
         setInternalIsOpen(false);
       }
     };
-
-    document.addEventListener('pointerdown', handleAction);
-    document.addEventListener('mousedown', handleAction);
-    return () => {
-      document.removeEventListener('pointerdown', handleAction);
-      document.removeEventListener('mousedown', handleAction);
-    };
+    document.addEventListener('mousedown', handleClickOutside);
+    return () => document.removeEventListener('mousedown', handleClickOutside);
   }, [internalIsOpen, closeOnOverlayClick]);
 
   // Handle focus management
   useEffect(() => {
     if (internalIsOpen) {
-      // Store the currently focused element
       previousFocusRef.current = document.activeElement as HTMLElement;
-
-      // Focus popover after it's rendered
-      if (popoverRef.current) {
-        popoverRef.current.focus();
-      }
+      popoverRef.current?.focus();
     } else {
-      // Return focus to previously focused element
-      if (previousFocusRef.current) {
-        previousFocusRef.current.focus();
-        previousFocusRef.current = null;
-      }
+      previousFocusRef.current?.focus();
+      previousFocusRef.current = null;
     }
   }, [internalIsOpen]);
 
+  const offsetStyle = useMemo(() => {
+    switch (effectivePlacement) {
+      case 'top':
+        return { marginBottom: `${offset}px` };
+      case 'bottom':
+        return { marginTop: `${offset}px` };
+      case 'left':
+        return { marginRight: `${offset}px` };
+      case 'right':
+        return { marginLeft: `${offset}px` };
+      default:
+        return {};
+    }
+  }, [effectivePlacement, offset]);
+
+  // Trigger cloning
   const triggerProps = trigger.props as {
     ref?: React.Ref<HTMLElement>;
     onClick?: (e: React.MouseEvent<HTMLElement>) => void;
   };
-  const triggerElement = React.cloneElement(
-    trigger as React.ReactElement,
-    {
-      'aria-expanded': internalIsOpen,
-      'aria-haspopup': 'dialog',
-      'aria-controls': popoverId,
-      ref: mergeRefs(triggerRef, triggerProps.ref),
-      onClick: (e: React.MouseEvent<HTMLElement>) => {
-        if (triggerProps.onClick) {
-          triggerProps.onClick(e);
-        }
-
-        // If clicking the trigger should not close the popover when it's open
-        if (!closeOnTriggerClick && internalIsOpen) {
-          return;
-        }
-
-        // Uncontrolled mode
-        if (isOpen === undefined) {
-          if (e.defaultPrevented) return;
-          if (popoverRef.current?.contains(e.target as Node)) return;
-          setInternalIsOpen((prev) => !prev);
-        }
-      },
-    } as Record<string, unknown>
-  );
+  const triggerElement = React.cloneElement(trigger, {
+    'aria-expanded': internalIsOpen,
+    'aria-haspopup': 'dialog',
+    'aria-controls': popoverId,
+    ref: mergeRefs(triggerRef, triggerProps.ref),
+    onClick: (e: React.MouseEvent<HTMLElement>) => {
+      triggerProps.onClick?.(e);
+      if (!closeOnTriggerClick && internalIsOpen) return;
+      if (isOpen === undefined) {
+        if (e.defaultPrevented) return;
+        if (popoverRef.current?.contains(e.target as Node)) return;
+        setInternalIsOpen((prev) => !prev);
+      }
+    },
+  } as Record<string, unknown>);
 
   return (
     <div id={id} ref={ref} className='relative inline-block'>
@@ -141,11 +139,12 @@ export function Popover({
         id={popoverId}
         ref={popoverRef}
         className={join(
-          'bg-popover text-popover-foreground z-[90] absolute top-full mt-2 transform rounded-md shadow-lg transition-all ease-out',
+          'absolute z-[90] transform rounded-md shadow-lg bg-popover text-popover-foreground transition-all ease-out',
           internalIsOpen ? 'opacity-100 scale-100' : 'opacity-0 scale-75 pointer-events-none',
-          POPOVER_ALIGNMENT_CLASSES[alignment],
+          placementVariants[effectivePlacement][alignment],
           className
         )}
+        style={offsetStyle}
         role='dialog'
         aria-modal='true'
         tabIndex={-1}
